Move ImageId::GetCatalogue() out of Drawing.cpp
