+++ src/openrct2/drawing/Drawing.cpp
@@ -374,32 +374,6 @@ const TranslucentWindowPalette kTranslucentWindowPalettes[kColourNumTotal] = {
 };
 // clang-format on
 
-ImageCatalogue ImageId::GetCatalogue() const
-{
-    auto index = GetIndex();
-    if (index >= SPR_TEMP_BEGIN && index < SPR_TEMP_END)
-    {
-        return ImageCatalogue::temporary;
-    }
-    if (index < SPR_RCTC_G1_END)
-    {
-        return ImageCatalogue::g1;
-    }
-    if (index < SPR_G2_END)
-    {
-        return ImageCatalogue::g2;
-    }
-    if (index < SPR_CSG_END)
-    {
-        return ImageCatalogue::csg;
-    }
-    if (index < SPR_IMAGE_LIST_END)
-    {
-        return ImageCatalogue::object;
-    }
-    return ImageCatalogue::unknown;
-}
-
 void GfxFilterPixel(RenderTarget& rt, const ScreenCoordsXY& coords, FilterPaletteID palette)
 {
     Rectangle::filter(rt, { coords, coords }, palette);
